Working obn projects menu

+++ database/migrations/2024_12_01_193752_create_client_contacts_table.php
@@ -11,26 +11,30 @@ class CreateClientContactsTable extends Migration
      *
      * @return void
      */
+
     public function up()
     {
-        Schema::create('client_contacts', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('company_id')->nullable();
-            $table->unsignedInteger('user_id');
-            $table->uuid('contact_id');
-            $table->string('contact_name');
-            // $table->string('phone')->nullable();
-            $table->string('email');
-            // $table->string('title')->nullable();
-            $table->unsignedInteger('added_by')->nullable();
-            $table->unsignedInteger('last_updated_by')->nullable();
-            $table->timestamps();
+        if (!Schema::hasTable('client_contacts')) {
+            Schema::create('client_contacts', function (Blueprint $table) {
+                $table->increments('id');
+                $table->unsignedInteger('company_id')->nullable();
+                $table->unsignedInteger('user_id');
+                $table->uuid('contact_id');
+                $table->string('contact_name');
+                // $table->string('phone')->nullable();
+                $table->string('email');
+                // $table->string('title')->nullable();
+                $table->unsignedInteger('added_by')->nullable();
+                $table->unsignedInteger('last_updated_by')->nullable();
+                $table->timestamps();
 
-            $table->foreign('added_by', 'client_contacts_added_by_foreign')->references('id')->on('users')->onDelete('set NULL')->onUpdate('cascade');
-            $table->foreign('company_id', 'client_contacts_company_id_foreign')->references('id')->on('companies')->onDelete('cascade')->onUpdate('cascade');
-            $table->foreign('last_updated_by', 'client_contacts_last_updated_by_foreign')->references('id')->on('users')->onDelete('set NULL')->onUpdate('cascade');
-            $table->foreign('user_id', 'client_contacts_user_id_foreign')->references('id')->on('users')->onDelete('cascade')->onUpdate('cascade');
-        });
+                $table->foreign('added_by', 'client_contacts_added_by_foreign')->references('id')->on('users')->onDelete('set NULL')->onUpdate('cascade');
+                $table->foreign('company_id', 'client_contacts_company_id_foreign')->references('id')->on('companies')->onDelete('cascade')->onUpdate('cascade');
+                $table->foreign('last_updated_by', 'client_contacts_last_updated_by_foreign')->references('id')->on('users')->onDelete('set NULL')->onUpdate('cascade');
+                $table->foreign('user_id', 'client_contacts_user_id_foreign')->references('id')->on('users')->onDelete('cascade')->onUpdate('cascade');
+            });
+
+        }
     }
 
     /**
